requirements.txt
Flask==2.3.3
gTTS==2.5.4
torch==2.3.1
openai-whisper==20250625
googletrans==4.0.0-rc1
